Refactor crd functions
