+++ crd/register_test.go
@@ -1,0 +1,11 @@
+package crd
+
+import (
+	. "github.com/onsi/ginkgo"
+	//. "github.com/onsi/gomega"
+	//. "github.com/grtl/mysql-operator/crd/util"
+)
+
+var _ = Describe("Register", func() {
+
+})
